+++ Automate/Framework/CommandHandler.cs
@@ -79,66 +79,87 @@ namespace Pathoschild.Stardew.Automate.Framework
 
             // settings
             report.AppendLine("Settings:\n------------------------------");
-            if (!this.Config.AutomateShippingBin)
-                report.AppendLine("   Shipping bin disabled");
             report.AppendLine($"   Automation interval: {this.Config.AutomationInterval}");
-            if (this.Config.MachinePriority.Any())
-            {
-                report.AppendLine("   Machine priorities:");
-                foreach (var pair in this.Config.MachinePriority.OrderByDescending(p => p.Value))
-                    report.AppendLine($"      {pair.Key}: {pair.Value}");
-            }
             if (this.Config.ConnectorNames.Any())
+                report.AppendLine($"   Connectors: {string.Join(", ", from name in this.Config.ConnectorNames orderby name select $"'{name}'")}");
+
+            // per-machine settings
             {
-                report.AppendLine("   Connectors:");
-                foreach (string name in this.Config.ConnectorNames.OrderBy(p => p))
-                    report.AppendLine($"      '{name}'");
+                StringBuilder perMachineReport = new StringBuilder();
+
+                foreach (var config in this.Config.MachineOverrides.OrderBy(p => p.Key, StringComparer.OrdinalIgnoreCase))
+                {
+                    string[] customSettings = config.Value
+                        .GetCustomSettings()
+                        .OrderBy(p => p.Key)
+                        .Select(p => $"{p.Key}={p.Value}")
+                        .ToArray();
+
+                    if (customSettings.Any())
+                        perMachineReport.AppendLine($"      {config.Key}: {string.Join(", ", customSettings)}");
+                }
+
+                if (perMachineReport.Length > 0)
+                {
+                    report.AppendLine("   Per-machine settings:");
+                    report.Append(perMachineReport);
+                }
             }
             report.AppendLine();
             report.AppendLine();
 
-            // summary
+            // machine summary
             report.AppendLine("Summary:\n------------------------------");
             {
-                MachineGroup[] allGroups = this.ActiveMachineGroups.SelectMany(p => p.Value).ToArray();
-                IAutomationFactory[] customFactories = this.Factory.GetFactories().Where(p => p.GetType() != typeof(AutomationFactory)).ToArray();
+                // machine groups
+                if (Context.IsWorldReady)
+                {
+                    MachineGroup[] allGroups = this.ActiveMachineGroups.SelectMany(p => p.Value).ToArray();
+                    report.AppendLine($"   Found {allGroups.Length} machine groups in {this.ActiveMachineGroups.Count} locations, containing {allGroups.Sum(p => p.Machines.Length)} automated machines connected to {allGroups.Sum(p => p.Containers.Length)} containers.");
+                }
+                else
+                    report.AppendLine("   No save loaded.");
 
-                report.AppendLine($"Found {allGroups.Length} machine groups in {this.ActiveMachineGroups.Count} locations, containing {allGroups.Sum(p => p.Machines.Length)} automated machines connected to {allGroups.Sum(p => p.Containers.Length)} containers.");
+                // custom machine factories
+                IAutomationFactory[] customFactories = this.Factory.GetFactories().Where(p => p.GetType() != typeof(AutomationFactory)).ToArray();
                 if (customFactories.Any())
-                    report.AppendLine($"Custom automation factories found: {string.Join(", ", customFactories.Select(p => p.GetType().FullName).OrderBy(p => p))}.");
+                    report.AppendLine($"   Custom automation factories found: {string.Join(", ", customFactories.Select(p => p.GetType().FullName).OrderBy(p => p))}.");
             }
             report.AppendLine();
             report.AppendLine();
 
             // machine groups
-            report.AppendLine("Automated machine groups:\n------------------------------");
-            if (this.ActiveMachineGroups.Any())
+            if (Context.IsWorldReady)
             {
-                foreach (GameLocation location in this.ActiveMachineGroups.Keys.OrderBy(p => $"{p.Name} ({p.NameOrUniqueName})", StringComparer.OrdinalIgnoreCase))
+                report.AppendLine("Automated machine groups:\n------------------------------");
+                if (this.ActiveMachineGroups.Any())
                 {
-                    MachineGroup[] machineGroups = this.ActiveMachineGroups[location];
-
-                    report.AppendLine($"   {location.Name}{(location.NameOrUniqueName != location.Name ? $" ({location.NameOrUniqueName})" : "")}:");
-
-                    foreach (MachineGroup group in machineGroups)
+                    foreach (GameLocation location in this.ActiveMachineGroups.Keys.OrderBy(p => $"{p.Name} ({p.NameOrUniqueName})", StringComparer.OrdinalIgnoreCase))
                     {
-                        var tile = group.Tiles[0];
+                        MachineGroup[] machineGroups = this.ActiveMachineGroups[location];
 
-                        report.AppendLine($"      Group at ({tile.X}, {tile.Y}):");
+                        report.AppendLine($"   {location.Name}{(location.NameOrUniqueName != location.Name ? $" ({location.NameOrUniqueName})" : "")}:");
 
-                        foreach (KeyValuePair<string, int> pair in group.Machines.GroupBy(p => p.MachineTypeID).ToDictionary(p => p.Key, p => p.Count()).OrderByDescending(p => p.Value))
-                            report.AppendLine($"          {pair.Value} x {pair.Key}");
+                        foreach (MachineGroup group in machineGroups)
+                        {
+                            var tile = group.Tiles[0];
 
-                        foreach (KeyValuePair<string, int> pair in group.Containers.GroupBy(p => p.Name).ToDictionary(p => p.Key, p => p.Count()).OrderByDescending(p => p.Value))
-                            report.AppendLine($"          {pair.Value} x {pair.Key}");
+                            report.AppendLine($"      Group at ({tile.X}, {tile.Y}):");
+
+                            foreach (KeyValuePair<string, int> pair in group.Machines.GroupBy(p => p.MachineTypeID).ToDictionary(p => p.Key, p => p.Count()).OrderByDescending(p => p.Value))
+                                report.AppendLine($"          {pair.Value} x {pair.Key}");
+
+                            foreach (KeyValuePair<string, int> pair in group.Containers.GroupBy(p => p.Name).ToDictionary(p => p.Key, p => p.Count()).OrderByDescending(p => p.Value))
+                                report.AppendLine($"          {pair.Value} x {pair.Key}");
+                            report.AppendLine();
+                        }
+
                         report.AppendLine();
                     }
-
-                    report.AppendLine();
                 }
+                else
+                    report.AppendLine("   No machines are currently automated.");
             }
-            else
-                report.AppendLine("   No machines are currently automated.");
 
             this.Monitor.Log(report.ToString(), LogLevel.Info);
         }